requirements.txt
alembic==0.9.1
click==6.7
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.6
MarkupSafe==1.0
pbr==1.10.0
psycopg2==2.7.1
python-dateutil==2.6.0
python-dotenv==0.6.4
python-editor==1.0.3
pytz==2017.2
requests==2.14.2
simplejson==3.10.0
six==1.10.0
SQLAlchemy==1.1.9
tap.py==2.1
typing==3.6.1
Werkzeug==0.12.2
zulip-beta==0.2.5
